# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go-Rocks is a tech playground with Go, Gin",
        "title": "Swagger Go-Rocks",
        "contact": {
            "name": "Lingga Edo",
            "url": "http://www.github.com/linggaaskaedo",
            "email": "lemp.otis@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/division": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoint for get division with param",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Division"
                ],
                "summary": "Get list of division based on query params",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "enum": [
                            "must-revalidate",
                            "must-db-revalidate"
                        ],
                        "type": "string",
                        "description": "Request cache control",
                        "name": "Cache-Control",
                        "in": "header"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Search by Division Public ID",
                        "name": "public_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by Division name. use % for wildcard search.",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "false",
                            "true"
                        ],
                        "type": "string",
                        "description": "Search by deleted status",
                        "name": "is_deleted",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "id",
                            "-id",
                            "name",
                            "-name",
                            "created_at",
                            "-created_at"
                        ],
                        "type": "string",
                        "default": "-created_at",
                        "description": "Sort result by these attributes",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": " ",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": " ",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPDivisionsResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoint for creating a new division",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Division"
                ],
                "summary": "Create division",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Payload create division",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.DivisionCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPDivisionResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    }
                }
            }
        },
        "/division/{divisionID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoint for get a division with ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Division"
                ],
                "summary": "Get division by ID",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Division ID",
                        "name": "divisionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPDivisionResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.DivisionDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "isDeleted": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dto.Meta": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string",
                    "x-order": "0"
                },
                "status_code": {
                    "type": "integer",
                    "x-order": "1"
                },
                "status": {
                    "type": "string",
                    "x-order": "2"
                },
                "message": {
                    "type": "string",
                    "x-order": "3"
                },
                "error": {
                    "type": "object",
                    "x-order": "4"
                },
                "timestamp": {
                    "type": "string",
                    "x-order": "5"
                }
            }
        },
...
        "entity.Pagination": {
            "type": "object",
            "properties": {
                "current_elements": {
                    "type": "integer"
                },
                "current_page": {
                    "type": "integer"
                },
                "cursor_end": {
                    "type": "string"
                },
                "cursor_start": {
                    "type": "string"
                },
                "sort_by": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "total_elements": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "rest.CreateDivisionData": {
            "type": "object",
            "properties": {
                "division": {
                    "$ref": "#/definitions/rest.DivisionDataPayload"
                }
            }
        },
...
        "rest.DivisionCreateRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/rest.CreateDivisionData"
                }
            }
        },
        "rest.DivisionData": {
            "type": "object",
            "properties": {
                "division": {
                    "$ref": "#/definitions/dto.DivisionDTO"
                }
            }
        },
        "rest.DivisionDataPayload": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "rest.DivisionsData": {
            "type": "object",
            "properties": {
                "divisions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DivisionDTO"
                    }
                }
            }
        },
...
        "rest.HTTPDivisionResp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/rest.DivisionData"
                },
                "metadata": {
                    "$ref": "#/definitions/dto.Meta"
                }
            }
        },
        "rest.HTTPDivisionsResp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/rest.DivisionsData"
                },
                "metadata": {
                    "$ref": "#/definitions/dto.Meta"
                },
                "pagination": {
                    "$ref": "#/definitions/entity.Pagination"
                }
            }
        },
...
        "rest.HTTPErrResp": {
            "type": "object",
            "properties": {
                "metadata": {
                    "$ref": "#/definitions/dto.Meta"
                }
            }
        },
...
    }
}